:memo: [Modify] KakaoServiceImpl 맴버 생성 코드 개선, 맴버 경력 조회, 삭제 기능 구현

+++ src/test/java/com/team3/itability/mypage/dao/MemberProfileDAOTest.java
@@ -23,7 +23,7 @@ class MemberProfileDAOTest {
     @Autowired
     ImageDAO imageDAO;
 
-    private int memberCode =1;
+    private Long memberCode =1L;
 
     @Test
     void name() {
